page: moved HomePage to its own file
